requirements.txt
pandas
numpy
pycryptodome